+++ tests/test_components.py
@@ -1,13 +1,14 @@
 import pytest
 from sympl import (
-    ConstantPrognostic, ConstantDiagnostic, RelaxationPrognostic, DataArray
+    ConstantPrognostic, ConstantDiagnostic, RelaxationPrognostic, DataArray,
+    timedelta,
 )
 import numpy as np
 
 
 def test_constant_prognostic_empty_dicts():
     prog = ConstantPrognostic({}, {})
-    tendencies, diagnostics = prog({})
+    tendencies, diagnostics = prog({'time': timedelta(0)})
     assert isinstance(tendencies, dict)
     assert isinstance(diagnostics, dict)
     assert len(tendencies) == 0
@@ -20,24 +21,81 @@ def test_constant_prognostic_cannot_modify_through_input_dict():
     prog = ConstantPrognostic(in_tendencies, in_diagnostics)
     in_tendencies['a'] = 'b'
     in_diagnostics['c'] = 'd'
-    tendencies, diagnostics = prog({})
+    tendencies, diagnostics = prog({'time': timedelta(0)})
     assert len(tendencies) == 0
     assert len(diagnostics) == 0
 
 
 def test_constant_prognostic_cannot_modify_through_output_dict():
     prog = ConstantPrognostic({}, {})
-    tendencies, diagnostics = prog({})
+    tendencies, diagnostics = prog({'time': timedelta(0)})
     tendencies['a'] = 'b'
     diagnostics['c'] = 'd'
-    tendencies, diagnostics = prog({})
+    tendencies, diagnostics = prog({'time': timedelta(0)})
     assert len(tendencies) == 0
     assert len(diagnostics) == 0
 
 
+def test_constant_prognostic_tendency_properties():
+    tendencies = {
+        'tend1': DataArray(
+            np.zeros([10]),
+            dims=['dim1'],
+            attrs={'units': 'm/s'},
+        ),
+        'tend2': DataArray(
+            np.zeros([2, 2]),
+            dims=['dim2', 'dim3'],
+            attrs={'units': 'degK/s'},
+        )
+    }
+    prog = ConstantPrognostic(tendencies)
+    assert prog.tendency_properties == {
+        'tend1': {
+            'dims': ('dim1',),
+            'units': 'm/s',
+        },
+        'tend2': {
+            'dims': ('dim2', 'dim3'),
+            'units': 'degK/s'
+        }
+    }
+    assert prog.diagnostic_properties == {}
+    assert prog.input_properties == {}
+
+
+def test_constant_prognostic_diagnostic_properties():
+    tendencies = {}
+    diagnostics = {
+        'diag1': DataArray(
+            np.zeros([10]),
+            dims=['dim1'],
+            attrs={'units': 'm'},
+        ),
+        'diag2': DataArray(
+            np.zeros([2, 2]),
+            dims=['dim2', 'dim3'],
+            attrs={'units': 'degK'},
+        )
+    }
+    prog = ConstantPrognostic(tendencies, diagnostics)
+    assert prog.diagnostic_properties == {
+        'diag1': {
+            'dims': ('dim1',),
+            'units': 'm',
+        },
+        'diag2': {
+            'dims': ('dim2', 'dim3'),
+            'units': 'degK',
+        }
+    }
+    assert prog.tendency_properties == {}
+    assert prog.input_properties == {}
+
+
 def test_constant_diagnostic_empty_dict():
     diag = ConstantDiagnostic({})
-    diagnostics = diag({})
+    diagnostics = diag({'time': timedelta(0)})
     assert isinstance(diagnostics, dict)
     assert len(diagnostics) == 0
 
@@ -46,22 +104,50 @@ def test_constant_diagnostic_cannot_modify_through_input_dict():
     in_diagnostics = {}
     diag = ConstantDiagnostic(in_diagnostics)
     in_diagnostics['a'] = 'b'
-    diagnostics = diag({})
+    diagnostics = diag({'time': timedelta(0)})
     assert isinstance(diagnostics, dict)
     assert len(diagnostics) == 0
 
 
 def test_constant_diagnostic_cannot_modify_through_output_dict():
     diag = ConstantDiagnostic({})
-    diagnostics = diag({})
+    diagnostics = diag({'time': timedelta(0)})
     diagnostics['c'] = 'd'
-    diagnostics = diag({})
+    diagnostics = diag({'time': timedelta(0)})
     assert len(diagnostics) == 0
 
 
+def test_constant_diagnostic_diagnostic_properties():
+    diagnostics = {
+        'diag1': DataArray(
+            np.zeros([10]),
+            dims=['dim1'],
+            attrs={'units': 'm'},
+        ),
+        'diag2': DataArray(
+            np.zeros([2, 2]),
+            dims=['dim2', 'dim3'],
+            attrs={'units': 'degK'},
+        )
+    }
+    diagnostic = ConstantDiagnostic(diagnostics)
+    assert diagnostic.diagnostic_properties == {
+        'diag1': {
+            'dims': ('dim1',),
+            'units': 'm',
+        },
+        'diag2': {
+            'dims': ('dim2', 'dim3'),
+            'units': 'degK',
+        }
+    }
+    assert diagnostic.input_properties == {}
+
+
 def test_relaxation_prognostic_at_equilibrium():
-    prognostic = RelaxationPrognostic('quantity')
+    prognostic = RelaxationPrognostic('quantity', 'degK')
     state = {
+        'time': timedelta(0),
         'quantity': DataArray(np.array([0., 1., 2.]), attrs={'units': 'degK'}),
         'quantity_relaxation_timescale': DataArray(
             np.array([1., 1., 1.]), attrs={'units': 's'}),
@@ -73,8 +159,9 @@ def test_relaxation_prognostic_at_equilibrium():
 
 
 def test_relaxation_prognostic_with_change():
-    prognostic = RelaxationPrognostic('quantity')
+    prognostic = RelaxationPrognostic('quantity', 'degK')
     state = {
+        'time': timedelta(0),
         'quantity': DataArray(np.array([0., 1., 2.]), attrs={'units': 'degK'}),
         'quantity_relaxation_timescale': DataArray(
             np.array([1., 1., 1.]), attrs={'units': 's'}),
@@ -86,8 +173,9 @@ def test_relaxation_prognostic_with_change():
 
 
 def test_relaxation_prognostic_with_change_different_timescale_units():
-    prognostic = RelaxationPrognostic('quantity')
+    prognostic = RelaxationPrognostic('quantity', 'degK')
     state = {
+        'time': timedelta(0),
         'quantity': DataArray(np.array([0., 1., 2.]), attrs={'units': 'degK'}),
         'quantity_relaxation_timescale': DataArray(
             np.array([1/60., 2/60., 3/60.]), attrs={'units': 'minutes'}),
@@ -99,8 +187,9 @@ def test_relaxation_prognostic_with_change_different_timescale_units():
 
 
 def test_relaxation_prognostic_with_change_different_equilibrium_units():
-    prognostic = RelaxationPrognostic('quantity')
+    prognostic = RelaxationPrognostic('quantity', 'm')
     state = {
+        'time': timedelta(0),
         'quantity': DataArray(np.array([0., 1., 2.]), attrs={'units': 'm'}),
         'quantity_relaxation_timescale': DataArray(
             np.array([1., 2., 3.]), attrs={'units': 's'}),
@@ -110,38 +199,6 @@ def test_relaxation_prognostic_with_change_different_equilibrium_units():
     tendencies, diagnostics = prognostic(state)
     assert np.all(tendencies['quantity'].values == np.array([1., 1., 1.]))
 
-
-def test_relaxation_prognostic_caching_timescale_and_equilibrium():
-    prognostic = RelaxationPrognostic(
-        'quantity',
-        relaxation_timescale=DataArray(
-            np.array([1., 1., 1.]), attrs={'units': 's'}),
-        equilibrium_value=DataArray(
-            np.array([1., 3., 5.]), attrs={'units': 'degK'}))
-    state = {
-        'quantity': DataArray(np.array([0., 1., 2.]), attrs={'units': 'degK'}),
-    }
-    tendencies, diagnostics = prognostic(state)
-    assert np.all(tendencies['quantity'].values == np.array([1., 2., 3.]))
-
-
-def test_relaxation_prognostic_cannot_override_cached_values():
-    prognostic = RelaxationPrognostic(
-        'quantity',
-        relaxation_timescale=DataArray(
-            np.array([1., 1., 1.]), attrs={'units': 's'}),
-        equilibrium_value=DataArray(
-            np.array([1., 3., 5.]), attrs={'units': 'degK'}))
-    state = {
-        'quantity': DataArray(np.array([0., 1., 2.]), attrs={'units': 'degK'}),
-        # random values which should not be used by the prognostic
-        'quantity_relaxation_timescale': DataArray(
-            np.array([5., 1., 18.]), attrs={'units': 's'}),
-        'equilibrium_quantity': DataArray(
-            np.array([4., 7., 2.])*1e-3, attrs={'units': 'km'}),
-    }
-    tendencies, diagnostics = prognostic(state)
-    assert np.all(tendencies['quantity'].values == np.array([1., 2., 3.]))
 
 if __name__ == '__main__':
     pytest.main([__file__])